added database of melodies and other stuff (see description)
Now we have some folders where we have midi files of several melodies.
Most of them are from bach, because the database where I searched the
melodies are mostly for classical music but other are also found.

Along with that I put in the requirements this: music21==3.2.1

and a todo.txt and a ideas.txt as proposed by oze.